refactor sections
