+++ app/components/sections/awardstemp.tsx
@@ -1,0 +1,5 @@
+export default function Awards() {
+  return (
+    <div>awards</div>
+  )
+}
